Added Score
I added a scoreboard that constantly updates every tick, rendering it on the screen. However, this does cause the game to be somewhat laggy, and I haven't figured out a good way to get around that. We could always just revert back to the old version without the text.

I also made it so that the player can't go up as high, but to keep the dash mechanic useful, I made it so that when the player dashes and they hit an enemy while dashing, they don't die.